+++ Debug/firmware/drivers/CoreI2C/i2c_interrupt.d
@@ -6,7 +6,11 @@ firmware/drivers/CoreI2C/i2c_interrupt.o: \
  ../firmware/drivers/CoreI2C/../../hal/hal_assert.h \
  ../firmware/drivers/CoreI2C/../../hal/../CMSIS/mss_assert.h \
  ../firmware/drivers/CoreI2C/core_i2c.h \
- ../firmware/drivers/CoreI2C/../../hal/CortexM3/GNU/cpu_types.h
+ ../firmware/drivers/CoreI2C/../../hal/CortexM3/GNU/cpu_types.h \
+ ../firmware/drivers/CoreI2C/../../CMSIS/m2sxxx.h \
+ C:/Microsemi/SoftConsole_v4.0/CMSIS/V4.3/Include/core_cm3.h \
+ C:/Microsemi/SoftConsole_v4.0/CMSIS/V4.3/Include/core_cmInstr.h \
+ C:/Microsemi/SoftConsole_v4.0/CMSIS/V4.3/Include/core_cmFunc.h
 
 ../firmware/drivers/CoreI2C/../../hal/hal.h:
 
@@ -21,3 +25,11 @@ firmware/drivers/CoreI2C/i2c_interrupt.o: \
 ../firmware/drivers/CoreI2C/core_i2c.h:
 
 ../firmware/drivers/CoreI2C/../../hal/CortexM3/GNU/cpu_types.h:
+
+../firmware/drivers/CoreI2C/../../CMSIS/m2sxxx.h:
+
+C:/Microsemi/SoftConsole_v4.0/CMSIS/V4.3/Include/core_cm3.h:
+
+C:/Microsemi/SoftConsole_v4.0/CMSIS/V4.3/Include/core_cmInstr.h:
+
+C:/Microsemi/SoftConsole_v4.0/CMSIS/V4.3/Include/core_cmFunc.h: