+++ inc/MPU6050/mpu6050.h
@@ -256,7 +256,7 @@ Mpu6050_Config_t Mpu6050_Config_;
 MPU6050_t Mpu6050_;
 
 /* ----- coreI2C instance ----- */
-i2c_instance_t g_i2c_inst_0;
+i2c_instance_t *g_core_i2c0;
 
 /* ========== Functions ========== */
 /** ------------------ Mpu_Init() ------------------